algo_pipeline_executer/requirements.txt
google-cloud-storage==2.11.0
google-cloud-aiplatform==1.24.0
numpy==1.24.3
google-auth==2.17.3
python-dotenv==1.0.0
charset-normalizer==3.2.0
MarkupSafe==2.1.3
google-cloud-run==0.10.5
requests==2.31.0
google-cloud-firestore==2.16.0
opencv-python-headless==4.9.0.80
pillow==10.3.0